+++ src/WeihanLi.Npoi/NpoiHelper.cs
@@ -21,24 +21,15 @@ internal static class NpoiHelper
             ? sheetSettings[sheetIndex]
             : sheetSettings[0];
 
-    /// <summary>
-    ///     Import sheet data to entity list
-    /// </summary>
-    /// <typeparam name="TEntity">entity type</typeparam>
-    /// <param name="sheet">excel sheet</param>
-    /// <param name="sheetIndex">sheetIndex</param>
-    /// <param name="dataAction">data action</param>
-    /// <returns>entity list</returns>
-    public static List<TEntity?> SheetToEntityList<TEntity>(ISheet? sheet, int sheetIndex, Action<TEntity?, ExcelConfiguration<TEntity>, int>? dataAction = null) where TEntity : new()
+    public static IEnumerable<TEntity?> SheetToEntities<TEntity>(ISheet? sheet, int sheetIndex, Action<TEntity?, ExcelConfiguration<TEntity>, int>? dataAction = null) where TEntity : new()
     {
         if (sheet is null || sheet.PhysicalNumberOfRows <= 0)
         {
-            return new List<TEntity?>();
+            yield break;
         }
 
         var configuration = InternalHelper.GetExcelConfigurationMapping<TEntity>();
         var sheetSetting = GetSheetSetting(configuration.SheetSettings, sheetIndex);
-        var entities = new List<TEntity?>(sheet.LastRowNum - sheetSetting.HeaderRowIndex);
 
         var propertyColumnDictionary = InternalHelper.GetPropertyColumnDictionary(configuration);
         var propertyColumnDic = sheetSetting.HeaderRowIndex >= 0
@@ -78,7 +69,8 @@ internal static class NpoiHelper
                             continue;
                         }
 
-                        var col = propertyColumnDic.GetPropertySetting(row.GetCell(i).StringCellValue.Trim());
+                        var title = row.GetCell(i).StringCellValue.Trim();
+                        var col = propertyColumnDic.GetPropertySetting(title);
                         if (null != col)
                         {
                             col.ColumnIndex = i;
@@ -87,7 +79,7 @@ internal static class NpoiHelper
                 }
 
                 // use default column index if no headers
-                if (propertyColumnDic.Values.All(_ => _.ColumnIndex < 0))
+                if (propertyColumnDic.Values.Any(_ => _.ColumnIndex < 0))
                 {
                     propertyColumnDic = propertyColumnDictionary;
                 }
@@ -101,7 +93,7 @@ internal static class NpoiHelper
 
                 if (row is null)
                 {
-                    entities.Add(default);
+                    yield return default;
                 }
                 else
                 {
@@ -160,17 +152,17 @@ internal static class NpoiHelper
                         }
                     }
 
-                    if (configuration.DataFilter?.Invoke(entity) != false)
+                    if (configuration.DataFilter?.Invoke(entity) == false)
                     {
-                        entities.Add(entity);
+                        continue;
                     }
 
                     dataAction?.Invoke(entity, configuration, rowIndex);
+
+                    yield return entity;
                 }
             }
         }
-
-        return entities;
     }
 
     private static void ProcessImport(object entity, IRow row, int rowIndex,
@@ -203,28 +195,36 @@ internal static class NpoiHelper
                         else
                         {
                             var valueApplied = false;
-                            if (InternalCache.ColumnInputFormatterFuncCache.TryGetValue(key,
-                                out var formatterFunc) && formatterFunc?.Method != null)
+                            InternalCache.CellReaderFuncCache.TryGetValue(key, out var cellReader);
+                            if (cellReader?.Method != null)
                             {
-                                var cellValue = cell.GetCellValue<string>(formulaEvaluator);
-                                try
+                                columnValue = cellReader.DynamicInvoke(cell);
+                                valueApplied = true;
+                            }
+                            else
+                            {
+                                InternalCache.ColumnInputFormatterFuncCache.TryGetValue(key,
+                                    out var formatterFunc);
+                                if (formatterFunc?.Method != null)
                                 {
-                                    // apply custom formatterFunc
-                                    columnValue = formatterFunc.DynamicInvoke(cellValue);
-                                    valueApplied = true;
-                                }
-                                catch (Exception e)
-                                {
-                                    Debug.WriteLine(e);
-                                    InvokeHelper.OnInvokeException?.Invoke(e);
+                                    var cellValue = cell.GetCellValue<string>(formulaEvaluator);
+                                    try
+                                    {
+                                        // apply custom formatterFunc
+                                        columnValue = formatterFunc.DynamicInvoke(cellValue);
+                                        valueApplied = true;
+                                    }
+                                    catch (Exception e)
+                                    {
+                                        Debug.WriteLine(e);
+                                        InvokeHelper.OnInvokeException?.Invoke(e);
+                                    }
                                 }
                             }
-
                             if (valueApplied == false)
                             {
                                 columnValue = cell.GetCellValue(key.PropertyType, formulaEvaluator);
                             }
-
                             valueSetter.Invoke(entity, columnValue);
                         }
                     }
@@ -241,7 +241,7 @@ internal static class NpoiHelper
     /// <param name="entityList">entity list</param>
     /// <param name="sheetIndex">sheetIndex</param>
     /// <returns>sheet</returns>
-    public static ISheet EntityListToSheet<TEntity>(ISheet sheet, IEnumerable<TEntity>? entityList, int sheetIndex)
+    public static ISheet EntitiesToSheet<TEntity>(ISheet sheet, IEnumerable<TEntity>? entityList, int sheetIndex)
     {
         Guard.NotNull(sheet, nameof(sheet));
         if (entityList is null)